+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] / 2022-02-15
+- Add GitVersion.CommandLine
+- Remove PackageDownload GitVersion.Tool
+- Add GitVersion.Tool
+- Include="NuGet.CommandLine"
+- Set Nuke.Common as *
+- Remove ValueInjectionUtility
+- Remove Logger to Serilog.Log
+- Update to Nuke.Common 6.0.0 Version
+
 ## [1.0.2] / 2022-01-19
 - Fix Version with 3 fields